story-drafting/infra/requirements.txt
aws-cdk-lib==2.233.0
constructs>=10.0.0,<11.0.0
tr-cdk-lib==2.16.1
